+++ app/profile/[userCid]/myboard/page.tsx
